architecture-fundamentals/requirements.txt

fastapi
uvicorn
kafka-python
pydantic
